# Conflicts:
#	.developer-vars.yml
#	.gitlab-ci.yml
#	health-mall-product-webapp/src/main/java/com/drstrong/health/product/ProductSpringBootApplication.java
#	health-mall-product-webapp/src/main/resources/application.yml

+++ health-mall-product-model/src/main/java/com/drstrong/health/product/model/response/store/StoreInfoResponse.java
@@ -31,7 +31,7 @@ public class StoreInfoResponse implements Serializable {
 	private Integer storeStatus;
 
 	@ApiModelProperty("店铺 sku 商品的数量")
-	private Long skuCount;
+	private Integer skuCount;
 
 	@ApiModelProperty("店铺的包邮额度")
 	private BigDecimal freePostage;
